requirements.txt
openai
beautifulsoup4
pydub
python-dotenv
